pyproject.toml
[project]
name = "pr-bot-azdo"
version = "1.0.0"
description = "A PR Review bot for Azure DevOps"
readme = "README.md"
authors = [
    { name = "Youri Immerzeel"}
]
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.19.0",
    "databricks-sdk>=0.62.0",
    "fastapi-throttle>=0.1.8",
    "fastapi[standard]>=0.116.1",
    "fastmcp==2.13.0",
    "hatchling>=1.27.0",
    "httpx>=0.28.1",
    "logfire[fastapi]>=4.3.3",
    "loguru>=0.7.3",
    "openai>=1.106.1",
    "pre-commit>=4.3.0",
    "pydantic-ai==1.15.0",
    "pydantic-settings>=2.6.1",
    "pyjwt>=2.10.1",
    "pyright>=1.1.403",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.12.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["app"]


[tool.ruff]
line-length = 120
lint.ignore = [
#    "E501", # line-too-long
    "D203", #incorrect-blank-line-before-class
    "D213" #multi-line-summary-second-line
]
lint.extend-select = ["D"] #PEP 257 docstring compliance
extend-exclude = [".databricks/.*", ".vscode/*", "tests/*"]

[tool.ruff.lint.per-file-ignores]
"!src/**.py" = ["D"] # Only demand docstring checks on src, too harsh to do on tests as well

#[tool.ruff.lint.pydocstyle]
#convention = "google"   #Deactivated for now

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["tests"] #pyright on tests would be too harsh
